Fix Bug
Bug:
the json serializer doesn't cast to List<int> but just parses it as List<dynamic> and because List<int> isn't a subtype of List<dynamic> we need to cast it ourselves (alternatively you could expect to receive List<dynamic>)

https://github.com/google/json_serializable.dart/issues/652

+++ lib/util/trends.dart
@@ -20,8 +20,8 @@ class Trends {
     return new Trends(
         userAnswer: json['user']['answer'],
         cpuAnswer: json['cpu']['answer'],
-        userWeeklyScores: json['user']['weeklyScores'],
-        cpuWeeklyScores: json['cpu']['weeklyScores']);
+        userWeeklyScores: json['user']['weeklyScores'].cast<int>(),
+        cpuWeeklyScores: json['cpu']['weeklyScores'].cast<int>());
   }
 }
